+++ cmake_format/test/test_lineend_win.cmake
@@ -1,0 +1,5 @@
+#[[*********************************************
+* Information line 1
+* Information line 2
+************************************************]]
+cmake_minimum_required(VERSION 2.8.11)